Refactor TemplateEngine and ClusterInstance packages
This commit refactors the TemplateEngine used to render install templates. The following changes are made:
- Refactor TemplateEngine into an interface type defined under internal/templates/engine.
- Define a concrete ClusterInstance TemplateEngine type that implements the interface.
- Refactor the clusterinstance package by separating the ClusterData into its own source file.
- Refactor the clusterinstance package by moving commonly used functions and constants to internal/controller/common package.
- Reorganized rendered_object.go source file (along with its unit tests) to internal/templates/engine/.
- Renamed the imagebasedinstall package to imagebasedinstaller.
- Added Red Hat copyright headers to internal/controller/conditions and internal/controller/retry source files.
