+++ internal/controller/clusterinstance/cluster_data.go
@@ -19,19 +19,13 @@ package clusterinstance
 import (
 	"encoding/json"
 	"fmt"
-	"html/template"
 	"strings"
 
-	sprig "github.com/go-task/slim-sprig"
 	"github.com/stolostron/siteconfig/api/v1alpha1"
-	"k8s.io/apimachinery/pkg/types"
-	k8syaml "sigs.k8s.io/yaml"
 )
 
 const (
 	cpuPartitioningKey = "cpuPartitioningMode"
-	AnnotationsKey     = "annotations"
-	LabelsKey          = "labels"
 )
 
 type SpecialVars struct {
@@ -41,10 +35,52 @@ type SpecialVars struct {
 }
 
 // ClusterData is a special object that provides an interface to the ClusterInstance spec fields for use in rendering
-// templates
+// ClusterInstance-based install templates
 type ClusterData struct {
 	Spec        v1alpha1.ClusterInstanceSpec
 	SpecialVars SpecialVars
+}
+
+// BuildClusterData returns a Cluster object that is consumed for rendering ClusterInstance-based install templates
+func BuildClusterData(
+	clusterInstance *v1alpha1.ClusterInstance,
+	node *v1alpha1.NodeSpec,
+) (data *ClusterData, err error) {
+
+	// Prepare specialVars
+	var currentNode v1alpha1.NodeSpec
+	if node != nil {
+		currentNode = *node
+	}
+
+	installConfigOverrides, err := getInstallConfigOverrides(clusterInstance)
+	if err != nil {
+		installConfigOverrides = ""
+	}
+
+	// Determine the number of control-plane and worker agents
+	controlPlaneAgents := 0
+	workerAgents := 0
+	for _, node := range clusterInstance.Spec.Nodes {
+		switch node.Role {
+		case "master":
+			controlPlaneAgents++
+		case "worker":
+			workerAgents++
+		}
+	}
+
+	data = &ClusterData{
+		Spec: clusterInstance.Spec,
+		SpecialVars: SpecialVars{
+			CurrentNode:            currentNode,
+			InstallConfigOverrides: installConfigOverrides,
+			ControlPlaneAgents:     controlPlaneAgents,
+			WorkerAgents:           workerAgents,
+		},
+	}
+
+	return
 }
 
 // getWorkloadPinningInstallConfigOverrides applies workload pinning to install config overrides if applicable
@@ -123,73 +159,6 @@ func getInstallConfigOverrides(clusterInstance *v1alpha1.ClusterInstance) (strin
 	}
 }
 
-// buildClusterData returns a Cluster object that is consumed for rendering templates
-func buildClusterData(clusterInstance *v1alpha1.ClusterInstance, node *v1alpha1.NodeSpec) (data *ClusterData, err error) {
-
-	// Prepare specialVars
-	var currentNode v1alpha1.NodeSpec
-	if node != nil {
-		currentNode = *node
-	}
-
-	installConfigOverrides, err := getInstallConfigOverrides(clusterInstance)
-	if err != nil {
-		installConfigOverrides = ""
-	}
-
-	// Determine the number of control-plane and worker agents
-	controlPlaneAgents := 0
-	workerAgents := 0
-	for _, node := range clusterInstance.Spec.Nodes {
-		switch node.Role {
-		case "master":
-			controlPlaneAgents++
-		case "worker":
-			workerAgents++
-		}
-	}
-
-	data = &ClusterData{
-		Spec: clusterInstance.Spec,
-		SpecialVars: SpecialVars{
-			CurrentNode:            currentNode,
-			InstallConfigOverrides: installConfigOverrides,
-			ControlPlaneAgents:     controlPlaneAgents,
-			WorkerAgents:           workerAgents,
-		},
-	}
-
-	return
-}
-
-// pruneManifest function returns true if the manifest should be pruned (i.e. not rendered)
-func pruneManifest(resource v1alpha1.ResourceRef, pruneList []v1alpha1.ResourceRef) bool {
-	if resource.APIVersion == "" || resource.Kind == "" || len(pruneList) == 0 {
-		return false
-	}
-
-	for _, p := range pruneList {
-		if resource.APIVersion == p.APIVersion && resource.Kind == p.Kind {
-			return true
-		}
-	}
-	return false
-}
-
-// suppressManifest function returns true if the manifest should be suppressed (i.e. not rendered)
-func suppressManifest(kind string, supressManifestsList []string) bool {
-	if kind == "" || len(supressManifestsList) == 0 {
-		return false
-	}
-
-	for _, manifest := range supressManifestsList {
-		if manifest == kind {
-			return true
-		}
-	}
-	return false
-}
-
 // mergeJSONCommonKey merge 2 json in common key and return string
 func mergeJSONCommonKey(mergeWith, mergeTo, key string) (string, error) {
 	var (
@@ -238,79 +207,4 @@ func mergeJSONCommonKey(mergeWith, mergeTo, key string) (string, error) {
 	}
 
 	return string(mergedJSON), nil
-}
-
-func appendToManifestMetadata(
-	appendData map[string]string,
-	field string,
-	manifest map[string]interface{},
-) map[string]interface{} {
-	if manifest["metadata"] == nil && len(appendData) > 0 {
-		manifest["metadata"] = make(map[string]interface{})
-	}
-	metadata, _ := manifest["metadata"].(map[string]interface{})
-
-	if metadata[field] == nil && len(appendData) > 0 {
-		metadata[field] = make(map[string]interface{})
-	}
-	data, _ := metadata[field].(map[string]interface{})
-
-	for key, value := range appendData {
-		if _, found := data[key]; !found {
-			// It's a new data-item, adding
-			if data == nil {
-				data = make(map[string]interface{})
-			}
-			data[key] = value
-		}
-	}
-	return manifest
-}
-
-func appendManifestAnnotations(extraAnnotations map[string]string, manifest map[string]interface{}) map[string]interface{} {
-	return appendToManifestMetadata(extraAnnotations, AnnotationsKey, manifest)
-}
-
-func appendManifestLabels(extraLabels map[string]string, manifest map[string]interface{}) map[string]interface{} {
-	return appendToManifestMetadata(extraLabels, LabelsKey, manifest)
-}
-
-// toYaml marshals a given field to Yaml
-func toYaml(v interface{}) string {
-	data, err := k8syaml.Marshal(v)
-	if err != nil {
-		// Swallow errors inside of a template.
-		return ""
-	}
-	return strings.TrimSuffix(string(data), "\n")
-}
-
-// funcMap provides additional useful functions for template rendering
-func funcMap() template.FuncMap {
-	f := sprig.TxtFuncMap()
-	f["toYaml"] = toYaml
-	return f
-}
-
-// GenerateOwnedByLabelValue is a utility function that generates the ownedBy label value
-// using the ClusterInstance namespace and name
-func GenerateOwnedByLabelValue(namespace, name string) string {
-	return fmt.Sprintf("%s_%s", namespace, name)
-}
-
-// GetNamespacedNameFromOwnedByLabel extracts the namespace and name from the ownedBy label value
-func GetNamespacedNameFromOwnedByLabel(ownedByLabel string) (types.NamespacedName, error) {
-	res := strings.Split(ownedByLabel, "_")
-	if len(res) != 2 {
-		return types.NamespacedName{}, fmt.Errorf("expecting single underscore delimiter in %s label value", ownedByLabel)
-	}
-
-	return types.NamespacedName{Namespace: res[0], Name: res[1]}, nil
-}
-
-func GetResourceId(name, namespace, kind string) string {
-	if namespace != "" {
-		return fmt.Sprintf("%s:%s/%s", kind, namespace, name)
-	}
-	return fmt.Sprintf("%s:%s", kind, name)
 }
