+++ internal/templates/engine/template_engine_interface.go
@@ -1,0 +1,34 @@
+/*
+Copyright 2024.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package engine
+
+import (
+	"context"
+
+	"go.uber.org/zap"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+// TemplateEngine defines an interface for processing templates for an abritary object
+type TemplateEngine interface {
+	ProcessTemplates(
+		ctx context.Context,
+		c client.Client,
+		log *zap.Logger,
+		object interface{},
+	) (RenderedObjectCollection, error)
+}
