+++ internal/templates/engine/test_utils.go
@@ -1,0 +1,38 @@
+/*
+Copyright 2024.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package engine
+
+import (
+	"reflect"
+	"sort"
+)
+
+func renderedObjectsEqual(objects1, objects2 []RenderedObject) bool {
+	if len(objects1) != len(objects2) {
+		return false
+	}
+
+	sort.Slice(objects1, func(x, y int) bool {
+		return objects1[x].GetKind() < objects1[y].GetKind()
+	})
+
+	sort.Slice(objects2, func(x, y int) bool {
+		return objects2[x].GetKind() < objects2[y].GetKind()
+	})
+
+	return reflect.DeepEqual(objects1, objects2)
+}
