+++ internal/templates/image-based-installer/template.go
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 */
 
-package imagebasedinstall
+package imagebasedinstaller
 
 const ImageClusterInstall = `apiVersion: extensions.hive.openshift.io/v1alpha1
 kind: ImageClusterInstall
